Refactor app structure, fix dependencies, and secure API keys
- Updated `react-router-dom` to `^6.22.0` and `dotenv` to `^16.4.5` in `package.json`.
- Refactored `src/ui/App.jsx` to implement client-side routing and authentication flow.
- Created `src/ui/Dashboard.jsx` for main application logic and `src/ui/ProtectedRoute.jsx` for access control.
- Fixed syntax error in `src/ui/Login.jsx` (import hoisting).
- Fixed memory leaks in `src/nlp/cd-detector.js` using `tf.tidy()` and tensor disposal.
- Enabled JWT interceptor in `src/services/api.js`.
- Implemented backend proxy for Gemini API calls in `server/index.js` to secure API keys.
- Refactored `src/nlp/reframe-engine.js` to use backend `/api/reframe` endpoint.
- Updated `server/index.js` to serve React app correctly in production (SPA routing).
- Removed client-side Gemini key usage.
