+++ package.json
@@ -20,14 +20,14 @@
     "bcryptjs": "^3.0.3",
     "body-parser": "^2.2.1",
     "cors": "^2.8.5",
-    "dotenv": "^17.2.3",
+    "dotenv": "^16.4.5",
     "express": "^5.2.1",
     "gemini-mcp-tool": "^1.1.4",
     "gemini-zod": "^0.1.2",
     "jsonwebtoken": "^9.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^7.11.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
     "styled-components": "^6.1.1"
   },
