+++ src/nlp/reframe-engine.js
@@ -1,34 +1,21 @@
-// src/nlp/reframe-engine.js
-import { GoogleGenerativeAI } from '@google/generative-ai';
-
-// Access your API key as an environment variable (preferable for security).
-// Make sure to set GEMINI_API_KEY in your environment variables.
-const API_KEY = process.env.GEMINI_API_KEY;
-
-if (!API_KEY) {
-  console.error("GEMINI_API_KEY environment variable not set. Reframing suggestions will not be generated.");
-}
-
-const genAI = API_KEY ? new GoogleGenerativeAI(API_KEY) : null;
-const model = genAI ? genAI.getGenerativeModel({ model: "gemini-pro" }) : null;
+import apiClient from '../services/api';
 
 async function getReframe(distortionType, originalText) {
-  if (!model) {
-    return 'Try to identify and rephrase unhelpful thinking. (AI model not loaded)';
-  }
-
-  const prompt = `The user expressed: "${originalText}"
-They are exhibiting a "${distortionType}" cognitive distortion.
-Please provide a concise and helpful reframing suggestion for this thought.`;
-
   try {
-    const result = await model.generateContent(prompt);
-    const response = await result.response;
-    const text = response.text();
-    return text.trim();
+    const response = await apiClient.post('/reframe', {
+      distortionType,
+      originalText
+    });
+    return response.data.suggestion;
   } catch (error) {
-    console.error("Error generating reframing suggestion with Gemini API:", error);
-    return 'Could not generate reframing suggestion at this time. Please try again later.';
+    console.error("Error generating reframing suggestion:", error);
+    if (error.response) {
+      return error.response.data.error || 'Could not generate reframing suggestion. Please try again.';
+    } else if (error.request) {
+      return 'No response from server. Please check your connection.';
+    } else {
+      return 'An error occurred while requesting reframe.';
+    }
   }
 }
